// packages/frontend/src/generated/types/DispatchMeasurementCompositeRequest.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export interface DispatchMeasurementCompositeRequest {
  /**
   * @type string,null
   */
  afterDensity?: string | null
  /**
   * @type string,null
   */
  afterHeight?: string | null
  /**
   * @type string
   */
  afterMass: string
  /**
   * @type string,null
   */
  afterVolume?: string | null
  /**
   * @type string,null
   */
  beforeDensity?: string | null
  /**
   * @type string,null
   */
  beforeHeight?: string | null
  /**
   * @type string
   */
  beforeMass: string
  /**
   * @type string,null
   */
  beforeVolume?: string | null
  /**
   * @type string, uuid
   */
  storageId: string
}