// packages/frontend/src/generated/zod/dispatchItemCompositeRequestSchema.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { ToZod } from '../.kubb/ToZod.ts'
import type { DispatchItemCompositeRequest } from '../types/DispatchItemCompositeRequest.ts'
import * as z from 'zod'

export const dispatchItemCompositeRequestSchema = z.object({
  dispatchedAmount: z.string(),
  productId: z.string().uuid(),
  storageId: z.string().uuid(),
}) as unknown as ToZod<DispatchItemCompositeRequest>